Fix missing kcp=>contrib/kcp renames
